started live updates * generated list_of_emails tables, that contains - an email associated - a lot of mail_infos (or users) associated.
*the schema Email received a has_one :list_of_emails, in this case
to be created, an email needs to be put inside a map with :email tag,
and be created as a ListOfEmails.
-  a ListOfEmails received an belongs_to :email, and be processed with
a cast_assoc.
In other words, ListOfEmails and Email are created together.

* the schema EmailInfo received a belongs_to list_of emails, that need
be passed at creation with
the other fields, so it will be processed with a put_assoc.
 - the schema ListOfEmails receive a has_many :email_infos, doesnt
 need more to do.

Other files as the htmls were generated and will be modified later.